requirements.txt
flask
youtube-transcript-api
gunicorn
google-genai
